User page with logout y delete user complete.

+++ composeApp/src/commonMain/kotlin/com/redfrogec/credigo/data/model/User.kt
@@ -5,16 +5,16 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class User(
-    val id: Int,
+    val id: Long,
     val image: String?,
     val email: String,
     val name: String,
     val passwordHash: String,
-    val questionId: Int,
+    val questionId: Long,
     val response: String,
     val registerDate: LocalDateTime,
     val updateDate: LocalDateTime?,
-    val tokenId: Int?,
+    val tokenId: Long?,
     val token: String?,
     val tokenExpire: LocalDateTime?,
 )
