+++ composeApp/src/commonMain/kotlin/com/redfrogec/credigo/ui/viewModel/SignUpViewModel.kt
@@ -52,8 +52,8 @@ class SignUpViewModel(private val sdk: UserSDK) : ViewModel() {
     private val _securityQuestion = MutableStateFlow("")
     val securityQuestion: StateFlow<String> = _securityQuestion.asStateFlow()
 
-    private val _idSecurityQuestion = MutableStateFlow(0)
-    val idSecurityQuestion: StateFlow<Int> = _idSecurityQuestion.asStateFlow()
+    private val _idSecurityQuestion = MutableStateFlow(0L)
+    val idSecurityQuestion: StateFlow<Long> = _idSecurityQuestion.asStateFlow()
 
     private val _answer = MutableStateFlow("")
     val answer: StateFlow<String> = _answer.asStateFlow()
@@ -92,7 +92,7 @@ class SignUpViewModel(private val sdk: UserSDK) : ViewModel() {
         validateForm()
     }
 
-    fun onSecurityQuestionChanged(id: Int, question: String) {
+    fun onSecurityQuestionChanged(id: Long, question: String) {
         _idSecurityQuestion.value = id
         _securityQuestion.value = question
         validateForm()
